Split CI tests into a raw and evaluation section
* Restructure data by grouping commits by Pull Request
